Remove superfluous casts
-  Every non-const pointer converts automatically to void *.
-  Every pointer converts automatically to void *.
-  void * converts to any other pointer.
-  const void * converts to any other const pointer.
-  Integer variables convert to each other.

I changed the declaration of a few variables in order to allow removing
a cast.

However, I didn't attempt to edit casts inside comparisons, since they
are very delicate.  I also kept casts in variadic functions, since they
are necessary, and in allocation functions, because I have other plans
for them.

I also changed a few casts to int that are better as ptrdiff_t.

This change has triggered some warnings about const correctness issues,
which have also been fixed in this patch (see for example src/login.c).

+++ lib/commonio.c
@@ -216,8 +216,8 @@ int do_fcntl_lock (const char *file, bool log, short type)
 	fd = open (file, O_WRONLY, 0600);
 	if (-1 == fd) {
 		if (log) {
-				(void) fprintf (shadow_logfd, "%s: %s: %s\n",
-				                shadow_progname, file, strerror (errno));
+			(void) fprintf (shadow_logfd, "%s: %s: %s\n",
+			                shadow_progname, file, strerror (errno));
 		}
 		return 0;
 	}
@@ -518,7 +518,7 @@ int commonio_open (struct commonio_db *db, int mode)
 		goto cleanup_ENOMEM;
 	}
 
-	while (db->ops->fgets (buf, (int) buflen, db->fp) == buf) {
+	while (db->ops->fgets (buf, buflen, db->fp) == buf) {
 		while (   ((cp = strrchr (buf, '\n')) == NULL)
 		       && (feof (db->fp) == 0)) {
 			size_t len;
