+++ libsubid/api.c
@@ -66,12 +66,12 @@ int get_subid_owner(unsigned long id, enum subid_type id_type, uid_t **owner)
 
 int subid_get_uid_owners(uid_t uid, uid_t **owner)
 {
-	return get_subid_owner((unsigned long)uid, ID_TYPE_UID, owner);
+	return get_subid_owner(uid, ID_TYPE_UID, owner);
 }
 
 int subid_get_gid_owners(gid_t gid, uid_t **owner)
 {
-	return get_subid_owner((unsigned long)gid, ID_TYPE_GID, owner);
+	return get_subid_owner(gid, ID_TYPE_GID, owner);
 }
 
 static
